A save() method that finally works!
Run the KingdominoApplication.java to test creating the file. Note that you need to change the directory of which the game file is loaded. (Or not use the example load file provided by the prof and use your own way to populate the game)
You also need to change the saving directory to make sure such a saving path is possible.